requirements.txt
# --- requirements.txt (MODIFIED) ---
Flask
python-dotenv
firebase-admin
Flask-Cors
gunicorn
uvicorn
openai
celery
redis